// projects/cli/src/commands/showcase/schemas/showcase/schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ShowcaseConfig {
  /**
   * Name of generated application.
   */
  name?: string;
  /**
   * Repository url.
   */
  repositoryUrl?: string;
  /**
   * Glob used to find features.
   */
  featureGlob: string;
  /**
   * Glob used to ignore features.
   */
  featureIgnoreGlob?: string;
  /**
   * Whether to generate API documentation using index.ts files.
   */
  api?: boolean;
  /**
   * Whether to generate readme documentation using README.md files.
   */
  readme?: boolean;
  /**
   * Where to write the default application.
   */
  appOut?: string;
  /**
   * External links in the generated app.
   */
  appExternalLinks?: {
    name?: string;
    href?: string;
    [k: string]: unknown;
  }[];
  /**
   * Where to write components.
   */
  componentOut?: string;
  /**
   * Tsconfig path to use when generating the api
   */
  tsconfig?: string;
  [k: string]: unknown;
}